moved scale converter into proper directory
