+++ playground/src/parameters/scale-converters/LinearBipolar70DbScaleConverter.h
